refactor: unify services
