--- internal/services/post/create.go
@@ -1,7 +1,0 @@
-package post
-
-import "github.com/nix-united/golang-echo-boilerplate/internal/models"
-
-func (postService *Service) Create(post *models.Post) {
-	postService.DB.Create(post)
-}
