--- internal/services/post/update.go
@@ -1,12 +1,0 @@
-package post
-
-import (
-	"github.com/nix-united/golang-echo-boilerplate/internal/models"
-	"github.com/nix-united/golang-echo-boilerplate/internal/requests"
-)
-
-func (postService *Service) Update(post *models.Post, updatePostRequest *requests.UpdatePostRequest) {
-	post.Content = updatePostRequest.Content
-	post.Title = updatePostRequest.Title
-	postService.DB.Save(post)
-}
