--- internal/services/token/create_refresh_token.go
@@ -1,25 +1,0 @@
-package token
-
-import (
-	"time"
-
-	"github.com/nix-united/golang-echo-boilerplate/internal/models"
-
-	"github.com/golang-jwt/jwt/v5"
-)
-
-func (tokenService *Service) CreateRefreshToken(user *models.User) (t string, err error) {
-	claimsRefresh := &JwtCustomRefreshClaims{
-		ID: user.ID,
-		RegisteredClaims: jwt.RegisteredClaims{
-			ExpiresAt: jwt.NewNumericDate(time.Now().Add(time.Hour * ExpireRefreshCount)),
-		},
-	}
-	refreshToken := jwt.NewWithClaims(jwt.SigningMethodHS256, claimsRefresh)
-
-	rt, err := refreshToken.SignedString([]byte(tokenService.config.Auth.RefreshSecret))
-	if err != nil {
-		return "", err
-	}
-	return rt, err
-}
