--- internal/services/user/register.go
@@ -1,27 +1,0 @@
-package user
-
-import (
-	"github.com/nix-united/golang-echo-boilerplate/internal/requests"
-	"github.com/nix-united/golang-echo-boilerplate/internal/server/builders"
-
-	"golang.org/x/crypto/bcrypt"
-)
-
-func (userService *Service) Register(request *requests.RegisterRequest) error {
-	encryptedPassword, err := bcrypt.GenerateFromPassword(
-		[]byte(request.Password),
-		bcrypt.DefaultCost,
-	)
-	if err != nil {
-		return err
-	}
-
-	user := builders.NewUserBuilder().SetEmail(request.Email).
-		SetName(request.Name).
-		SetPassword(string(encryptedPassword)).
-		Build()
-
-	userService.DB.Create(&user)
-
-	return nil
-}
