+++ internal/services/user/service.go
@@ -2,7 +2,9 @@ package user
 
 import (
 	"github.com/nix-united/golang-echo-boilerplate/internal/requests"
+	"github.com/nix-united/golang-echo-boilerplate/internal/server/builders"
 
+	"golang.org/x/crypto/bcrypt"
 	"gorm.io/gorm"
 )
 
@@ -16,4 +18,23 @@ type Service struct {
 
 func NewUserService(db *gorm.DB) *Service {
 	return &Service{DB: db}
+}
+
+func (userService *Service) Register(request *requests.RegisterRequest) error {
+	encryptedPassword, err := bcrypt.GenerateFromPassword(
+		[]byte(request.Password),
+		bcrypt.DefaultCost,
+	)
+	if err != nil {
+		return err
+	}
+
+	user := builders.NewUserBuilder().SetEmail(request.Email).
+		SetName(request.Name).
+		SetPassword(string(encryptedPassword)).
+		Build()
+
+	userService.DB.Create(&user)
+
+	return nil
 }
